Introduce WorkflowContainer for running a workflow inside a Compose app.
This is the third flavor of integration, it replaces `setContentWorkflow`,
`WorkflowLayout`, `WorkflowRunnerViewModel`, etc., and makes it really
easy to run a `Workflow` inside a pure Compose app. It's compatible with
`ViewEnvironment`/`ViewRegistry`, but doesn't require it – it gives you the
root rendering, and you can do whatever you want with it.

It also supports running root `ComposeWorkflow`s directly, since they are
self-rendering.