Inline WorkflowViewStub logic in ViewFactory.showRendering to avoid unnecessary intermediate Views.
This is effectively the logic of `WorkflowViewStub`, but translated into Compose idioms.
This approach has a few advantages:

 - Avoids extra custom views required to host `WorkflowViewStub` inside a Composition. Its trick
   of replacing itself in its parent doesn't play nice with Compose.
 - Allows us to pass the correct parent view for inflation (the root of the composition).
 - Avoids `WorkflowViewStub` having to do its own lookup to find the correct [ViewFactory], since
   we already have the correct one.

Like `WorkflowViewStub`, this function uses the `ViewFactory` to create and memoize a `View` to
display the rendering, keeps it updated with the latest `RenderingT` and `ViewEnvironment`, and
adds it to the composition.