Renames ComposeViewFactoryRoot to CompositionRoot and decouples the implementation.
The root is now applied via a `ViewRegistry` wrapper that wraps individual factories
to apply the root, instead of having this logic hard-coded inside `ComposeViewFactory`.

This approach is more flexible in general (could be used to do other tricks), and decouples
the rooting feature from the rest of the code.