Make renderAsState public, make WorkflowContainer take a ViewEnvironment.
This narrows the scope of `WorkflowContainer` to be only for showing a workflow's renderings
using a `ViewEnvironment`. This elimnates the need to have separate overloads for workflows
with rendering type `ComposeRendering`, and is more idiomatic – the old `WorkflowContainer`
wasn't a "container" in any Compose-y sense of the word, and it was really weird that it took
a content function. Now this composable is actually a container for all workflow-related
composables, and so I think the name is actually appropriate (closes #22).

To render a workflow without a `ViewEnvironment`, `renderAsState` is now public. This is also
much more idiomatic, as it resembles APIs like `Flow<T>.collectAsState` and
`Observable.subscribeAsState`.