Fix subcomposition of ComposableViewFactory and WorkflowRendering.
Fixes to how subcomposition works after discussing with Leland.

Currently, there are two cases for calling `ViewGroup.setContent`:
1. **Root composition (no parent `CompositionReference`).** In this case, compose
   will throw if called multiple times, so we only call it once and then use
   a `MutableState` to feed rendering updates to it.
2. **Subcomposition.** In this case, we _must_ call `setContent` again (i.e.
   manually recompose the child composition) every time the composition that
   the `CompositionReference` belongs to is recomposed. This is not documented
   anywhere, but is apparently part of the contract. In this case, we call
   `setContent` every time we get a rendering update.

This also pulls the `bindView` call back into a custom view. I'm not entirely sure
why, but there seems to be a timing issue with initializing all the composables
and views when the logic is all done inside compose. Using a custom view also means
we don't need to access `OwnerAmbient`.

**Known issues:**

 - #43: This issue did not occur with dev11, so I think it's a Compose bug. I haven't
   officially filed yet since distilling a reproducer is probably going to be a lot
   of work.