Migrate WorkflowTester to renderWorkflowIn.
This exposed a few actual bugs in `RenderIdempotencyChecker`, the fact that
we still need `workerContext` for `renderWorkflowIn` (until GUWT workers are implemented),
and a bug in the sample `TimeMachineWorkflow` machinery.

It also introduced an explicit behavior change for `WorkflowTester`: `sendProps`
no longer triggers new render passes when props are sent that are equivalent to
the previous props. This is expected, since props is now a `StateFlow`, which implies
`distinctUntilChanged`-like behavior. This only required fixing one test that
was using the old behavior as a trick to trigger a new render pass with the same
state. If this affects actual tests, they can be fixed by wrapping in a workflow
that accepts some arbitrary changing value, and sending new props that change that
value instead of sending the same underlying props again.