Rename EmittedOutput to WorkflowOutput, move to workflow-core.
This simplifies the API of `WorkflowAction.applyTo`, removes the need for
`OutputHolder` in the core tests and `MaybeOutput` in `workflow-runtime`.
`EmittedOutput` is left as a deprecated typealias to ease migration.

I was initially reluctant to do something like this because I didn't want
Workflow to provide yet another generic `Optional` type. I forgot that the
testing library was already doing that though, and I think that using a
nullable "holder" type is a gross enough API that it won't likely be used
outside of the intended uses.