Remove unnecessary data classes from public, non-experimental APIs.
The only such data classes were `WorkflowTestParams` and `RenderingAndSnapshot`.
The former has no need to be a data class, as it's just passed as configuration
to the test function, and the latter is effectively just an alias for `Pair` and
the shape is fundamental to the workflow runtime design so it won't change.

There are still some data classes in the UI modules, but #55 will mark all those
as `ExperimentalWorkflowApi` before releasing 1.0.0.

Fixes #57.