Make RenderTester throw a more realistic exception if the same workflow/side effect is ran twice.
It now throws an exception that looks just like the real runtime. This doesn't _fail_ the test,
it _errors_ it, which is also better behavior since this case means the workflow-under-test is
written incorrectly, not that it doesn't match expected test behavior. Fixes #120.

This change also cleans up the way `WorkflowIdentifier`s are described in error and testing strings,
unifies the impostor and non-impostor calls, makes the string consistent between runtime and tests,
and removes a method from the public API.