Introduces `AndroidViewRendering`, de-emphasizes `ViewRegistry`
`AndroidViewRendering` is implemented by renderings that can provide their own
`ViewFactory`. `ViewRegistry.buildView` looks to see if a given rendering
implements that method if it fails to find an entry for it -- note that we
look for a registration first, so `ViewRegistry` is able override the default.

Given this, it's no longer crucial for developers to define and provide a
`ViewRegistry`, so we get rid of all the nannying that forced them to provide
one.

Updates the samples to use the new interface and drop their `ViewRegistry`
instances where possible.