Fix the double showRendering call when initializing DecorativeViewFactory views.
Introduces `ViewInitializer` which gets executed by `bindShowRendering` between setting
the tag and actually invoking `showRendering`, which gives the `DecorativeViewFactory`
logic the chance to cancel the original `showRendering` call since it wraps it.

This initializer concept will also be used in a follow-up PR to allow `ViewStateCache`
to correctly initialize AndroidX view tree owners.