Make ModalContainer look for the parent lifecycle using ViewTreeLifecycleOwner.
This stops the leak of #314, but doesn't implement the full fix agreed upon there.

This change also changes the API a bit so it doesn't pollute the entire auto-complete
namespace for `Context` (or `View`) for consumers, which is important because I'm about
to add some more functions there and Workflow isn't an "AndroidX support support library".