Improve support for intercepting RenderContext calls.
Instead of passing `BaseRenderContext` back to its `proceed` function,
`WorkflowInterceptor.onRender` now can provide an implementation of  a new
interface, `RenderContextInterceptor`. This is similar to
`WorkflowInterceptor`, with a bunch of methods that mirror
`BaseRenderContext`, each with its own `proceed` function. This makes it much
simpler to intercept calls to `actionSink.send` and to intercept side effects.

Also uses the new class in `SimpleLoggingWorkflowInterceptor` to add additional logging.

Fixes #330 and fixes #331.

Co-authored-by: Ray Ryan <ray@squareup.com>