(Re)introduce ViewTreeLifecycleOwner support to containers
Introduces something called `WorkflowLifecycleOwner` that implements `LifecycleOwner` and is set as the `ViewTreeLifecycleOwner` on the child views of the navigation containers: `WorkflowViewStub`, `BackstackContainer`, and `ModalContainer`. These containers all move their lifecycle to the `DESTROYED` state when they are about to discard the current child view and replace it with a new one because the rendering type changed. Until then, their lifecycle follows that of their parent lifecycle.

This PR was created in March, then left for a few months, and finally picked back up in August. That said, very little had to be changed. The updates to AndroidX and Kotlin were merged in separate PRs (#459, #460). It was started by reverting commit 8500b0670e52d02a7904a91dda7914b57a62b968 (which reverted 74e14e9c and cdca31ce).