Cleans up view initialization in DecorativeViewFactory.
```kotlin
initView: (OuterRenderingT, ViewEnvironment) -> Unit
```

is now

```kotlin
initializeView: View() -> Unit
```

just like the param on `ViewRegistry.buildView`. We also eliminate
the hard coded call to `showRendering()`, giving authors complete control
over when / if it's called. Again, this is in parallel with `ViewRegistry`'s
behavior.

`View.showFirstRendering()` is introduced, to simplify the 99% case.

Closes #396