Don't try to restore unrecognized view state.
Defangs a few dangerous calls to super from `View.onRestoreInstanceState`.
We'd crash in cases where the app was saved while showing a view of one type,
and then got restored showing another. Such a mismatch is reasonable --
imagine a configuration change that bounces between split pane and single, or
being saved on a logged in screen and restored after the session expires.

Such calls often accidentally work, if you follow the usual idiom of extending
`View.BaseSavedState`. But that idiom is not a requirement, and there's no
reason to believe it would be safe for a view of type A to try to restore the
state written by a view of type B.

It would be better to simply skip the call to super in such cases, but that's
not allowed. Instead we make a no-op call:

```
?: super.onRestoreInstanceState(super.onSaveInstanceState())
```

I've extended `BackStackContainerTest` to cover this, but we need new tests
for the other calls. Will add those in follow up.