Introduce support for ViewTreeSavedStateRegistryOwner in BackStackContainer.
This PR introduces the ability for BackStackContainer to manage `SavedStateRegistryOwner`s on the subviews it creates for navigation frames. Every immediate child view gets a registry owner, and that owner is set on the child view via `ViewTreeSavedStateRegistryOwner`. Any views in the subtree underneath a `BackStackContainer` can then query `ViewTreeSavedStateRegistryOwner` for a `SavedStateRegistry` and use that registry to save and restore their state. The registry will be correctly saved when something else is pushed on the backstack on top of that view, and restored when the view is shown again.

This motivation for adding this support is to support `AbstractComposeView` in workflows. Compose uses the `SavedStateRegistry` to save internal state for things like scrollables.

This change builds on #381 which adds support for `ViewTreeLifecycleOwner`. It is not my first attempt at adding support for this: see #463, which superseded #284. This PR is one more step in completing #458.