Introduce ComposeViewFactory as a convenient way to define ViewFactories as composable functions.
First part of #458.

This code was mostly copy/pasted from the compose sub-project, with some additional polish. Some of the docs refer to things that don't exist yet, but will be coming in immediate follow-ups.

I'll bring in the samples once more pieces are merged, so I don't have to edit them to work with a more limited set of primitives.

Archeology: The first version of this was introduced in https://github.com/square/workflow/pull/703.