Introduces Wrapper
We make a lot of use of renderings that wrap others to add meta information or
behavior, but which need to delegate for UI construction. This is tricky.

* We frequently forget to implement `Compatible`
* Things like `NamedViewFactory` or tricky `DecorativeViewFactory` implementations
  have to be added to `ViewRegistry`. Besides being fussy, these are platform specific.
* Introspective tasks like logging wind up with fragile code to peek
  through known wrappers.

The Wrapper interface and supporting extension functions should provide a
platform neutral way around this, and I hope pave the way to move
`ViewRegistry`(or a replacement) from `core-android` to `core-common`.