Introduces ScreenOverlay and ModalScreenOverlayDialogFactory
`ScreenOverlay` marks an `Overlay` with a `Screen` defining its content.

`ModalScreenOverlayDialogFactory` can show a `ScreenOverlay` as an
`android.app.Dialog`, taking care of updating the content view and managing
the Back button. It's the much simpler replacement for `ModalViewContainer`.

`ModalScreenOverlayDialogFactory` works with `BodyAndModalsContainer`
and `LayeredDialogs` to perform two important tricks:

 - `updateBounds(Dialog, Rect)`, which allows dialogs to be restrictred to a subset
   of the screen. Implemented via the private `ModalArea` value in `ViewEnvironment`

 - Touch and keyboard events are blocked by views that are covered by
   dialogs managed by a `ModalScreenOverlayDialogFactory`. Same for
   dialogs with a lower z order. We use another private `ViewEnvironment`
   value for that, `CoveredByModal`.

In sample code, `PanelOverlay` replaces `PanelContainerScreen`, and the Tic
Tac Toe sample is updated to use the new hotness. I think the diff of the
sample code really highlights how much the new marker interfaces improve our
composition story.

That said, the `ScrimScreen` bit in the sample is a little rough. In the old
code we magically swizzled the scrim into place automatically. This PR
seems big enough already, so I'll follow up with a clean up that does
something similar and deletes all the deprecated sample code.

Fixes #259, #138, #99, #204, #314, #589