Introduces ScreenViewFactoryFinder
Fixes #594, which was about how impractical it is to wrap the `ViewRegistry`
interface, by introducing a higher level interface that's easy to wrap.

The problem is that the fundamental method is `getFactoryFor(KClass)`, but
lots of crucial behavior (e.g. `AndroidViewRendering` /
`AndroidScreenRendering`) is in the `getFactoryForRendering` extension
method. But if we change the fundamental method to be instance based instead
of type based, we bring back a lot of potential complexity to `ViewRegistry`
that the original type-based choice very intentionally restricted.

To have our cake and eat it too, we move the extension method to a new
`ViewEnvironment` service interface, `ScreenViewFactoryFinder`. Ta da,
totally customizable, with the defaults totally built in.