Tightens up the API related to ViewTreeSavedStateRegistryOwner.
`StateRegistryAggregator` is renamed `WorkflowSavedStateRegistryAggregator`, and now works only with `KeyedSavedStateRegistryOwner`, rather than any old `SavedStateRegistryOwner`. In fact, `WorkflowSavedStateRegistryAggregator` is the now only way to create instances of `KeyedSavedStateRegistryOwner`, which is no longer public API. This eliminates the possibility of forgetting to make the initial call to `restoreChildRegistryOwnerIfReady`.

One slightly worrisome change is that `StateRegistryAggregator` now keeps a map of all the `KeyedStateRegistryOwner` instances it creates, theoretically increasing the chances that we'll leak them. Prior to this they were 1:1 with the associated views, with the idea that they wouldn't need to be cleaned up that way. (Note how the new map allows us to eliminate `onWillSave` and `onRestored` from the constructor.)

In reality, even before this change we still had to be careful to always call `prune` or `save` as managed views were dropped, so we weren't avoiding any bookkeeping. And there was a decent amount of crucial copy/pasted boilerplate surrounding each use of `KeyedStateRegistryOwner`, which caused real bugs.