659: Collect and test baseline profiles
Add in Macrobenchmark Module to gather baseline profiles and benchmark impact on same scenario in the Dungeon Sample App.

Add the full profile the Dungeon sample app and add in the split profile sections to each module in the Workflow library.

X2b still not seeing performance improvements, but 10% improvement reported on Pixel. Will be testing on phone soon.