wip: AliasRendering : VisibleRendering
More ambitious alternative to #695, nowhere near working yet.

Note in particular that the resolved environment is lost on the first rendering.

I think the key to really solving all this, and making it as easy as it should be to wrap view and dialog factories, is to introduce some kind of wrapper for the resolved factory and resolution machinery. Basically, add the factory to `WorkflowViewState`, and make `WorkflowViewState` responsible for answering the `View.canShowRendering` question instead of comparing the new rendering to the current one the way we do now.

That would mean that `DecorativeScreenViewFactory` no longer has to replace the "real" rendering with the wrapper, and it would get a lot simpler. `WorkflowViewState` is what gets returned from `ViewEnvironment.unwrapRenderingAndGetFactory`, and is useful to both `ScreenViewFactory` and `OverlayDialogFactory`.

It'd be nice if this all got simpler instead of more complicated.