Fixes rapid BackStackContainer.update calls.
Turns out that `TransitionManager.go()` is a no-op if there is already a transition in progress, so a new screen shown before the previous one was done animating would never be displayed. Even better, a redundant call to display that last screen again would cause a crash in `WorkflowSavedStateRegistryAggregator`, when we would try to register the name already logged for the lost view.