LegacyComposeViewTreeIntegrationTest is more purely legacy.
I had cheated a bit on this, wrapping various things so that I could avoid the deprecated methods on WorkflowViewStub. Doing so caused testing difficulties in #703 that seem purely related to the tests themselves, not a reflection of problems in real life.

So we revert the wrapping in `LegacyComposeViewTreeIntegrationTest` and give it its own `Activity` to run against, full of comfortaing deprecation warnings.