Defers most deprecation and restores deleted kdoc
I think we should hold off on deprecating the old API for the first few pre-release drops, to give established code bases (like Square's) a chance to catch up before being flooded with warnings. Note that I've commented out the `@Deprecation` and `@Suppress` lines rather than deleting them, so that they'll be easier to bring back.

We keep the deprecations on `WorkflowLayout` and `WorkflowViewStub`, though, because they're so easy to honor.